core/requirements.txt
psycopg2-binary==2.8.6
python-dotenv==0.15.0
